+++ UnitTest/Tests/TestGameInfoSyncAction.cs
@@ -27,6 +27,44 @@ namespace UnitTest.Tests
         }
 
         [TestMethod]
+        public void ApplyToGameFile_FindsTitleInTheLastGAMEINFO()
+        {
+            var gameFile = new GameFile()
+            {
+                FileName = "blah.wad"
+            };
+
+            var syncAction = new GameInfoSyncAction();
+
+            Tree files = new Tree("root", 
+                new Tree("GAMEINFO", "STARTUPTITLE  = \"Wrong answer\""),
+                new Tree("GAMEINFO.real", "STARTUPTITLE  = \"Right answer\""));
+            var reader = new TreeReader(files);
+
+            var result = syncAction.ApplyToGameFile(gameFile, reader, new string[0]);
+
+            Assert.AreEqual("Right answer", gameFile.Title);
+        }
+
+        [TestMethod]
+        public void ApplyToGameFile_IgnoresSuperchargeTitleInGAMEINFO()
+        {
+            var gameFile = new GameFile()
+            {
+                FileName = "blah.wad"
+            };
+
+            var syncAction = new GameInfoSyncAction();
+
+            Tree files = new Tree("root", new Tree("GAMEINFO", "IWAD = \"blah.wad\"\nSTARTUPTITLE  = \"Supercharge\"\nSOMETHING_ELSE = blah"));
+            var reader = new TreeReader(files);
+
+            var result = syncAction.ApplyToGameFile(gameFile, reader, new string[0]);
+
+            Assert.IsTrue(string.IsNullOrEmpty(gameFile.Title));
+        }
+
+        [TestMethod]
         public void ApplyToGameFile_FindsIntendedGameInGAMEINFO()
         {
             var gameFile = new GameFile()